+++ poptorch/source/PoptorchSymbols.hpp
@@ -33,6 +33,7 @@ extern c10::Symbol numpy_T;                     // NOLINT
 extern c10::Symbol col2im;                      // NOLINT
 extern c10::Symbol im2col;                      // NOLINT
 extern c10::Symbol _reshape_alias;              // NOLINT
+extern c10::Symbol argsort;                     // NOLINT
 
 // Removed in 1.10.
 #if TORCH_MINOR_VERSION < 10